moves create-store from utils/middleware to utils
